Replaces implementation with merging strategy.
This version of the tilemap merges the geometry of each
tile to reduce the number of GPU calls, but instead increases
the necessary amount of memory.

This allows us to reuse materials accurately and render
quickly, but takes up a lot of memory as we use a non-shared
geometry buffer for each tile instance.