Adds an implementation of instanced rendering.
This implementation is added alongside the merged
implementation and uses the InstancedBufferGeometry
feature alongside a custom vertex shader to render
tile instances at various locations.

This uses minimal memory (because the instances
are sharing geometry and using a buffer attribute)
and can still render quickly (because all instances
are handled within a draw call.

The downside is that it requires a non-trivial
customization of the vertex shader that may interfere
with material rendering properties, and also prevents
the scene graph from being able to run other geometric
operations such as ray-collision-queries because the
geometry only gets created via the instancing on the
GPU.